Refactor auth utils and add facility migration function.
