+++ kolibri/core/auth/utils/users.py
@@ -1,0 +1,9 @@
+from kolibri.core.auth.models import AdHocGroup
+from kolibri.core.auth.models import Membership
+
+
+def create_adhoc_group_for_learners(classroom, learners):
+    adhoc_group = AdHocGroup.objects.create(name="Ad hoc", parent=classroom)
+    for learner in learners:
+        Membership.objects.create(user=learner, collection=adhoc_group)
+    return adhoc_group
